chore: upgrade dependency
Snyk has created this PR to upgrade typescript from 5.2.2 to 5.3.2.

See this package in npm:
https://www.npmjs.com/package/typescript

See this project in Snyk:
https://app.snyk.io/org/jim60105-cSFDhUtg8KgwHDFgLvtMMX/project/362050f7-c9c0-4d63-92af-e6d5adb1d1a3?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

fix: upgrade eslint-plugin-vue from 9.17.0 to 9.19.2 (#6)

Snyk has created this PR to upgrade eslint-plugin-vue from 9.17.0 to 9.19.2.

See this package in npm:
https://www.npmjs.com/package/eslint-plugin-vue

See this project in Snyk:
https://app.snyk.io/org/jim60105-cSFDhUtg8KgwHDFgLvtMMX/project/362050f7-c9c0-4d63-92af-e6d5adb1d1a3?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

fix: upgrade vue from 3.3.4 to 3.3.9 (#5)

Snyk has created this PR to upgrade vue from 3.3.4 to 3.3.9.

See this package in npm:
https://www.npmjs.com/package/vue

See this project in Snyk:
https://app.snyk.io/org/jim60105-cSFDhUtg8KgwHDFgLvtMMX/project/362050f7-c9c0-4d63-92af-e6d5adb1d1a3?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

fix: upgrade primevue from 3.35.0 to 3.42.0 (#4)

Snyk has created this PR to upgrade primevue from 3.35.0 to 3.42.0.

See this package in npm:
https://www.npmjs.com/package/primevue

See this project in Snyk:
https://app.snyk.io/org/jim60105-cSFDhUtg8KgwHDFgLvtMMX/project/362050f7-c9c0-4d63-92af-e6d5adb1d1a3?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

fix: upgrade @types/node from 18.18.4 to 18.18.14 (#3)

Snyk has created this PR to upgrade @types/node from 18.18.4 to 18.18.14.

See this package in npm:
https://www.npmjs.com/package/@types/node

See this project in Snyk:
https://app.snyk.io/org/jim60105-cSFDhUtg8KgwHDFgLvtMMX/project/362050f7-c9c0-4d63-92af-e6d5adb1d1a3?utm_source=github&utm_medium=referral&page=upgrade-pr

Co-authored-by: snyk-bot <snyk-bot@snyk.io>

Upgrade lock file

Signed-off-by: 陳鈞 <jim60105@gmail.com>